[DEPENDENCIES] Removed Janet.
After doing a COMPLETE integration, it was giving weird compiler errors (trying to call the `true` constant like it was a function? And calling one of my types as a function with 4 args, even though there are no instances in the script of me trying to do that ...) I tried pulling latest version from the official repo and latest has compilation issues if the JANET_NET module is removed in janetconf.h. This was the last straw, the language isn't particularly great enough for me to continue this headache. I thought it would be a fun language to try and integrate, as I have been interested in toying with a LISP-like language, but Janet is not mature enough to be worth my time.

HOWEVER, I did get a lot of code cleanup and bug fixes out of this, so it wasn't a complete waste of time.