[SHIBBOLETH] Fixing compilation error caused by using static_cast instead of const_cast in InstancedArray.
Fixing compilation errors in RenderPipeline::getRenderStages().

ReflectionDefinition::base() will fall back to using ClassHashable if there is no reflection for a type. This reduces the need for the BASE() macro if a class has a ClassHashable template specialization.

Working on ModelComponent and the IModelStageRegistration interface for allowing entities with models to register with the appropriate render pipeline stages.